% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_packages.R
\name{base_packages}
\alias{base_packages}
\title{A list of all the base packages.
This is useful for knowing what not to try to install or uninstall.}
\usage{
base_packages()
}
\description{
A list of all the base packages.
This is useful for knowing what not to try to install or uninstall.
}
