% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deps.R
\name{get_deps}
\alias{get_deps}
\title{Get dependencies for a package.}
\usage{
get_deps(pkg, depends = TRUE, suggests = FALSE, recursive = FALSE,
  recursive_suggests = FALSE, error_on_cran = TRUE)
}
\arguments{
\item{pkg}{character. The name of the package to fetch dependencies for.}

\item{depends}{logical. Whether or not to fetch Imports / Depends dependencies. Defaults to
TRUE.}

\item{suggests}{logical. Whether or not to fetch Suggests depedencies. Defaults to FALSE.}

\item{recursive}{logical. Whether to recursively fetch Imports / Depends subdependencies for
all the fetched dependencies. Defaults to FALSE.}

\item{recursive_suggests}{logical. Whether to recursively fetch Suggests subdependices for
all the fetched dependencies. This is very slow! Defaults to FALSE.}

\item{error_on_cran}{logical. Whether to raise an error if the package is not on CRAN (e.g.,
it is a typo, or it is on Bioconductor). If FALSE, a warning will be raised instead.
Defaults to FALSE.}
}
\description{
Get dependencies for a package.
}
