% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deps_data.R
\name{get_deps_data}
\alias{get_deps_data}
\title{Get package data (name, version, license) for given dependencies.}
\usage{
get_deps_data(pkgs)
}
\arguments{
\item{pkgs}{character. A character vector of package names to fetch data for.}
}
\value{
a list of named lists containing package name, version, and license for each package.
}
\description{
Get package data (name, version, license) for given dependencies.
}
